refactor: MyInfo 모델을 매니저로 분리 (#99)
