+++ DiegLog/DiegLog/Model/Manager/ImageFileManager.swift
