+++ DiegLog/DiegLog/Model/Manager/RecentSearchManager.swift
