+++ DiegLog/DiegLog/Model/Manager/YoutubeAPI.swift
