+++ DiegLog/DiegLog/Model/Model/ExerciseModel.swift
